+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Now `paizaio.RunnersGetDetailsResponse` has `Error` field.
+- Better error handling.
 
 ## [0.5.1] - 2021-02-02